Renamed `engine_events.rs` file to `events.rs`
