+++ wolf_engine_core/src/events/mod.rs
@@ -5,5 +5,5 @@
 
 pub use generic_event_queue::*;
 
-mod engine_events;
-pub use engine_events::*;
+mod events;
+pub use events::*;
